requirements.txt
functions-framework==3.10.0
google-cloud-billing==1.18.0
Flask==3.1.2